pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = ""
authors = [
    {name = "Author", email = "author@example.com"},
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "transformers>=4.30.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "torch>=2.0.0",
    "plotly>=5.0.0",
    "tqdm>=4.60.0",
    "scipy>=1.9.0",
    "pillow>=9.0.0",
    "requests>=2.28.0",
    "open-clip-torch>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "syrupy>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
